Adds a bunch of KEVINTODO's where there is obvious work that
remains to be done, I'm sure it won't compile, all of the code
that I'd added in the defensive branch will still be using
int_function and int_block instead of the new names

Conflicts:
	dataflowAPI/src/SymEval.C
	dataflowAPI/src/stackanalysis.C
	dyninstAPI/h/BPatch_point.h
	dyninstAPI/src/BPatch_point.C
	dyninstAPI/src/BPatch_process.C
	dyninstAPI/src/Parsing.C
	dyninstAPI/src/Relocation/Atoms/Atom.C
	dyninstAPI/src/Relocation/Atoms/CFAtom.C
	dyninstAPI/src/Relocation/Transformers/Defensive.C
	dyninstAPI/src/addressSpace.C
	dyninstAPI/src/addressSpace.h
	dyninstAPI/src/baseTramp.C
	dyninstAPI/src/binaryEdit.C
	dyninstAPI/src/block.C
	dyninstAPI/src/emit-x86.C
	dyninstAPI/src/function.C
	dyninstAPI/src/hybridCallbacks.C
	dyninstAPI/src/hybridOverwrites.C
	dyninstAPI/src/inst-winnt.C
	dyninstAPI/src/instPoint.C
	dyninstAPI/src/instPoint.h
	dyninstAPI/src/mapped_object.C
	dyninstAPI/src/parse-cfg.C
	dyninstAPI/src/parse-cfg.h
	dyninstAPI/src/pdwinnt.C
	dyninstAPI/src/process.C
	dyninstAPI/src/registerSpace.C
	dyninstAPI/src/symtab.C
	dyninstAPI/src/symtab.h
	instructionAPI/src/Operation.C
	parseAPI/src/Function.C
	symtabAPI/src/emitWin.h

+++ stackwalk/src/linux-ppc-swk.C
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1996-2009 Barton P. Miller
+ * Copyright (c) 1996-2011 Barton P. Miller
  * 
  * We provide the Paradyn Parallel Performance Tools (below
  * described as "Paradyn") on an AS IS basis, and do not warrant its
@@ -50,7 +50,7 @@ unsigned ProcDebugLinux::getAddressWidth()
    return sizeof(void *);
 }
 
-#define USER_OFFSET_OF(register) ((signed int) &(((struct user *) NULL)->regs.register))
+#define USER_OFFSET_OF(register) ((signed long) &(((struct user *) NULL)->regs.register))
 
 static long int getRegOffset(Dyninst::MachRegister reg, int addr_width)
 {
